+++ ChangeLog.md
@@ -1,4 +1,5 @@
 #### v.4.0.7 (TBD)
+* Bug fix: DS VR type value ending with \0 fails to serialize in JsonDicomConverter (#1078)
 * Allow disabling validation in JsonDicomConverter
 
 #### v.4.0.6 (8/6/2020)